♻️ Refactor all "embeded" to "embed"
