+++ src/interfaces/Embed.ts
@@ -1,4 +1,4 @@
-export interface EmbededField {
+export interface embedField {
   name: string;
   value: string;
 }
